% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms1.R
\name{annotateMz}
\alias{annotateMz}
\title{Perform annotation on MS1 level}
\usage{
annotateMz(mz, compdb, adducts, tolerance = 0, ppm = 0)
}
\arguments{
\item{mz}{\code{numeric} neutral mass for which the adduct m/z shall be calculated.}

\item{compdb}{\code{data.frame} data frame containing compounds used for annoation.
Should minimally contain $name, $exactmass as input.}

\item{adducts}{\code{character} specifying the name of the adduct;
supported values are returned by \code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adductNames()}}}

\item{tolerance}{\code{numeric} absolute annotation error}

\item{ppm}{\code{numeric} ppm annotation error}
}
\value{
\code{list} List of data frames with the same length as mz.
}
\description{
\code{annotateMz} calculates the m/z value from a neutral mass and an adduct
definition.
}
\author{
Michael Witting
}
