% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSpectra.R
\name{CompareSpectraParam}
\alias{CompareSpectraParam}
\alias{CompareSpectraParam-class}
\alias{matchSpectra,Spectra,Spectra,CompareSpectraParam-method}
\title{Matching MS Spectra against a reference}
\usage{
CompareSpectraParam(
  MAPFUN = joinPeaks,
  tolerance = 0,
  ppm = 5,
  FUN = MsCoreUtils::ndotproduct,
  requirePrecursor = TRUE,
  requirePrecursorPeak = FALSE,
  THRESHFUN = function(x) which(x >= 0.7),
  ...
)

\S4method{matchSpectra}{Spectra,Spectra,CompareSpectraParam}(query, target, param, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{MAPFUN}{\code{function} used to map peaks between the compared spectra.
Defaults for \code{CompareSpectraParam} to \code{\link[=joinPeaks]{joinPeaks()}}. See
\code{\link[=compareSpectra]{compareSpectra()}} for details.}

\item{tolerance}{\code{numeric(1)} for an absolute maximal accepted difference
between m/z values. This will be used in \code{compareSpectra} as well as for
eventual precursor m/z matching.}

\item{ppm}{\code{numeric(1)} for a relative, m/z-dependent, maximal accepted
difference between m/z values. This will be used in \code{compareSpectra} as
well as for eventual precursor m/z matching.}

\item{FUN}{\code{function} used to calculate similarity between spectra. Defaults
for \code{CompareSpectraParam} to \code{\link[MsCoreUtils:distance]{MsCoreUtils::ndotproduct()}}. See
\code{\link[MsCoreUtils:distance]{MsCoreUtils::ndotproduct()}} for details.}

\item{requirePrecursor}{\code{logical(1)} whether only target spectra are
considered in the similarity calculation with a precursor m/z that matches
the precursor m/z of the query spectrum (considering also \code{ppm} and
\code{tolerance}). With \code{requirePrecursor = TRUE} (the default) the function
will complete much faster, but will not find any hits for target (or query
spectra) with missing precursor m/z.}

\item{requirePrecursorPeak}{\code{logical(1)} whether only target spectra will be
considered in the spectra similarity calculation that have a peak with an
m/z matching the precursor m/z of the query spectrum. Defaults to
\code{requirePrecursorPeak = FALSE}.}

\item{THRESHFUN}{\code{function} applied to the similarity score to define which
target spectra are considered \emph{matching}. Defaults to
\code{THRESHFUN = function(x) which(x >= 0.7)} hence selects
all target spectra matching a query spectrum with a similarity higher or
equal than \code{0.7}. Any function that takes a numeric vector with similarity
scores (as returned by \code{\link[=compareSpectra]{compareSpectra()}}) as input and returns a
\code{logical} or \code{integer} vector with the matches is supported.}

\item{...}{for \code{CompareSpectraParam}: additional parameters passed along
to the \code{\link[=compareSpectra]{compareSpectra()}} call.}

\item{query}{for \code{matchSpectra}: \link{Spectra} object with the query spectra.}

\item{target}{for \code{matchSpectra}: \link{Spectra} object with the target
(reference) spectra to compare \code{query} against.}

\item{param}{for \code{matchSpectra}: parameter object (such as
\code{CompareSpectraParam}) defining the settings for the matching.}

\item{BPPARAM}{for \code{matchSpectra}: parallel processing setup (see the
\code{BiocParallel} package for more information). By default parallel
processing is disabled.}
}
\description{
\code{matchSpectra} with both \code{query} and \code{target} being a \link{Spectra} object
matches each spectra in \code{query} against all spectra in \code{target} and reports
matches with a similarity that passes the \code{THRESHFUN} condition. The
parameters for the matching can be specified with one of the \code{param} objects
listed below:
\itemize{
\item \code{CompareSpectraParam}: the \emph{generic} parameter object allowing to set all
settings for the \code{\link[=compareSpectra]{compareSpectra()}} call that is used to perform the
similarity calculation. This includes \code{MAPFUN} and \code{FUN} defining the
peak-mapping and similarity calculation functions and \code{ppm} and \code{tolerance}
to define an acceptable difference between m/z values of the compared
peaks. Additional parameters to the \code{compareSpectra} call
can be passed along with \code{...}. See the help of \code{\link[=Spectra]{Spectra()}} for more
information on these parameters. Parameters \code{requirePrecursor} (default
\code{TRUE}) and \code{requirePrecursorPeak} (default \code{FALSE}) allow to pre-filter
the target spectra prior to the actual similarity calculation for each
individual query spectrum. This can considerably improve performance.
Finally, parameter \code{THRESHFUN} allows to define a function to be applied to
the similarity scores to define which matches to report. See below for more
details.
}
}
\author{
Johannes Rainer
}
