% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDbSource.R
\name{CompDbSource}
\alias{CompDbSource}
\alias{CompDbSource-class}
\alias{metadata,CompDbSource-method}
\alias{show,CompDbSource-method}
\title{Compound Annotation Sources for \code{CompDb} databases}
\usage{
CompDbSource(dbfile = character())

\S4method{metadata}{CompDbSource}(x, ...)

\S4method{show}{CompDbSource}(object)
}
\arguments{
\item{dbfile}{\code{character(1)} with the database file (including the full
path).}

\item{x}{A \code{CompDbSource} object.}

\item{...}{ignored.}

\item{object}{A \code{CompDbSource} object.}
}
\description{
\code{CompDbSource} objects represent references to \link{CompDb} database-backed
annotation resources. Instances are expected to be created with the dedicated
construction functions such as \code{MassBankSource} or the generic
\code{CompDbSource}. The annotation data is not stored within the object but will
be accessed/loaded within the object's \code{matchSpectra} method.

New \code{CompDbSource} objects can be created using the functions:
\itemize{
\item \code{CompDbSource}: create a new \code{CompDbSource} object from an existing
\code{CompDb} database. The (SQLite) database file (including the full path)
needs to be provided with parameter \code{dbfile}.
}

Other functions:
\itemize{
\item \code{metadata}: get metadata (information) on the annotation resource.
}
}
\examples{

## Locate a CompDb SQLite database file. For this example we use the test
## database from the `CompoundDb` package.
fl <- system.file("sql", "CompDb.MassBank.sql", package = "CompoundDb")
ann_src <- CompDbSource(fl)

## The object contains only the reference/link to the annotation resource.
ann_src
}
\author{
Johannes Rainer
}
