% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIonMode.R
\name{matchIonMode}
\alias{matchIonMode}
\alias{matchIonMode,numeric,numeric,IonModeParam-method}
\title{m/z matching}
\usage{
matchIonMode(query, target, param, ...)

\S4method{matchIonMode}{numeric,numeric,IonModeParam}(query, target, param, BPPARAM = SerialParam())
}
\description{
The \code{matchIonMode} function is a prototype.
}
