% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMatchedSpectra.R
\name{shinyMatchedSpectra}
\alias{shinyMatchedSpectra}
\title{Verify Results from matchSpectra}
\usage{
shinyMatchedSpectra(object)
}
\arguments{
\item{object}{A non-empty instance of class \code{MatchedSpectra}.}
}
\value{
A \code{MatchedSpectra} with verified results
}
\description{
The \code{shinyMatchedSpectra()} function opens a simple shiny application
that allows to browse results stored in a \code{MatchedSpectra} object. Results
can be verified and set to TRUE or FALSE. Upon pushing the "Save & Close"
button the the app is closed and a filtered \code{MatchedSpectra} is returned,
containing only results set to TRUE.
}
\author{
Carolin Huber, Michael Witting
}
