% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraDbSource.R
\name{SpectraDbSource}
\alias{SpectraDbSource}
\alias{SpectraDbSource-class}
\alias{WeizMassSource}
\alias{show,SpectraDbSource-method}
\title{Compond Annotation Sources for databases supporting \code{Spectra}}
\usage{
WeizMassSource(
  IagreeToTheLicense = FALSE,
  sqlite = FALSE,
  version = character(),
  user = character(),
  pass = character(),
  host = character(),
  dbname = character()
)

\S4method{show}{SpectraDbSource}(object)
}
\arguments{
\item{IagreeToTheLicense}{\code{logical(1)} specifying whether the user accepts
to the WeizMass license agreement or not.}

\item{version}{For \code{WeizMassSource}: \code{character(1)} specifying the version
(release) of the database to connect to.}

\item{user}{For \code{WeizMassSource}: optional \code{character(1)} with the
username to access the WeizMass database.}

\item{pass}{For \code{WeizMassSource}: optional \code{character(1)} with the
password to access the WeizMass database.}

\item{host}{For \code{WeizMassSource}: optional \code{character(1)} with the
name (address) of the host with the WeizMass database.}

\item{dbname}{For \code{WeizMassSource}: optional \code{character(1)} with the
name of the database on the server. If \code{sqlite = TRUE} the file name
of the SQLite database needs to be defined with this parameter.}

\item{object}{A \code{SpectraDbSource} object.}
}
\description{
\code{SpectraDbSource} objects represent references to annotation databases that
are supported by a dedicated \link{MsBackend} backend and hence allow to return
their data as a \link{Spectra} object. A specific \code{MsBackend} has thus to be
available (and needs to be supported) by the database. Examples are
\emph{MassBank} with the
\href{https://github.com/RforMassSpectrometry/MsBackendMassbank}{MsBackendMassbank}
backend or \emph{WeizMass} with the
\href{https://github.com/RforMassSpectrometry/MsBackendWeizMass}{MsBackendWeizMass}.
Additional parameters that need to be defined are the type of SQL database
system that is used by the resource (e.g. \code{MariaDB} for MySQL/MariaDB
servers) as well as credential information allowing to connect to the
database. Connection to the database is only performed in the \code{matchSpectra}
call and the returned \link{MatchedSpectra} object contains only matching spectra
from the reference (target) database.

New \code{SpectraDbSource} objects should not be created manually, but using one
of the constructor functions listed below. Functions are available for
different annotation databases that provide also useful defaults simplifying
access to these resources.
\itemize{
\item \code{WeizMassSource}: provides access to a WeizMass database. This database
can be a SQLite database (in which case the \code{sqlite} parameter has to be
set to \code{TRUE} and the database file has to be provided with the \code{dbname}
parameter) or a (remote) WeizMass database. In the latter case the version
of the resource has to be provided. Alternatively, using parameters \code{host},
\code{dbname}, \code{user} and \code{pass} it is possible to connect to a custom location
of a WeizMass database.
For WeizMass databases, users have to agree to the WeizMass license by
specifically setting parameter \code{IagreeToTheLicense} to \code{TRUE}.
}
}
\seealso{
Other Compound annotation sources: 
\code{\link{CompDbSource}}
}
\author{
Johannes Rainer, Nir Shachaf
}
\concept{Compound annotation sources}
