% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchedSummarizedExperiment.R,
%   R/hidden-aliases.R
\docType{methods}
\name{colnames,MatchedSummarizedExperiment-method}
\alias{colnames,MatchedSummarizedExperiment-method}
\alias{$,MatchedSummarizedExperiment-method}
\alias{matchedData,MatchedSummarizedExperiment-method}
\alias{keepMatches,MatchedSummarizedExperiment-method}
\alias{dropMatches,MatchedSummarizedExperiment-method}
\alias{hidden_aliases}
\title{Internal page for hidden aliases}
\usage{
\S4method{colnames}{MatchedSummarizedExperiment}(x)

\S4method{$}{MatchedSummarizedExperiment}(x, name)

\S4method{matchedData}{MatchedSummarizedExperiment}(object, columns = colnames(object), ...)

\S4method{keepMatches}{MatchedSummarizedExperiment}(
  object,
  queryValues = integer(),
  targetValues = integer(),
  queryColumnName = character(),
  targetColname = character(),
  idxs = integer(),
  ...
)

\S4method{dropMatches}{MatchedSummarizedExperiment}(
  object,
  queryValues = integer(),
  targetValues = integer(),
  queryColumnName = character(),
  targetColname = character(),
  idxs = integer(),
  ...
)
}
\value{
Not applicable
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
