% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden-aliases.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\title{Internal page for hidden aliases}
\value{
Not applicable
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
