% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFormula.R
\name{matchFormula}
\alias{matchFormula}
\alias{matchFormula,character,character-method}
\title{formula matching}
\usage{
matchFormula(query, target, ...)

\S4method{matchFormula}{character,character}(query, target, BPPARAM = SerialParam())
}
\arguments{
\item{query}{}

\item{target}{}

\item{...}{currently ignored}
}
\value{
\link{Matched} object representing the result
}
\description{
XXX
}
\examples{

XXX

}
\author{
Michael Witting
}
