% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMz.R
\name{matchMz}
\alias{matchMz}
\title{m/z matching}
\usage{
matchMz(x, y, adducts, tolerance, ppm)
}
\arguments{
\item{x}{Feature table containing information on MS1 features. Must contain
column called \code{mz}}

\item{y}{Compound table with metabolites to compare against. Must contain
columns called \code{exactmass} and \code{name}}

\item{adducts}{Allowed adducts named accordingly to \code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adductNames()}}}

\item{tolerance}{\code{numeric(1)} for an absolute maximal accepted difference
between m/z values.}

\item{ppm}{\code{numeric(1)} for a relative, m/z-dependent, maximal accepted
difference between m/z values.}
}
\description{
The \code{matchMz} method matches (compares) m/z values from a MS1 data table
with theoretical m/z values from compounds. \code{adducts} specifies the allowed
adducts based on adduct names from \code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adductNames()}}. The
maximum allowed error is defined by \code{tolerance} or \code{ppm}
}
\author{
Michael Witting
}
