% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMz.R
\name{MzParam}
\alias{MzParam}
\alias{Mass2MzParam}
\alias{Mass2MzRtParam}
\alias{MzRtParam}
\alias{matchMz}
\alias{matchMz,numeric,numeric,Mass2MzParam-method}
\alias{matchMz,numeric,data.frameOrSimilar,Mass2MzParam-method}
\alias{matchMz,data.frameOrSimilar,numeric,Mass2MzParam-method}
\alias{matchMz,data.frameOrSimilar,data.frameOrSimilar,Mass2MzParam-method}
\alias{matchMz,numeric,numeric,MzParam-method}
\alias{matchMz,numeric,data.frameOrSimilar,MzParam-method}
\alias{matchMz,data.frameOrSimilar,numeric,MzParam-method}
\alias{matchMz,data.frameOrSimilar,data.frameOrSimilar,MzParam-method}
\alias{matchMz,data.frameOrSimilar,data.frameOrSimilar,Mass2MzRtParam-method}
\alias{matchMz,data.frameOrSimilar,data.frameOrSimilar,MzRtParam-method}
\alias{matchMz,SummarizedExperiment,ANY,Param-method}
\title{m/z matching}
\usage{
MzParam(tolerance = 0, ppm = 5)

Mass2MzParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5)

Mass2MzRtParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5, toleranceRt = 0)

MzRtParam(tolerance = 0, ppm = 0, toleranceRt = 0)

matchMz(query, target, param, ...)

\S4method{matchMz}{numeric,numeric,Mass2MzParam}(query, target, param, BPPARAM = SerialParam())

\S4method{matchMz}{numeric,data.frameOrSimilar,Mass2MzParam}(
  query,
  target,
  param,
  massColname = "exactmass",
  BPPARAM = SerialParam()
)

\S4method{matchMz}{data.frameOrSimilar,numeric,Mass2MzParam}(query, target, param, BPPARAM = SerialParam(), mzColname = "mz")

\S4method{matchMz}{data.frameOrSimilar,data.frameOrSimilar,Mass2MzParam}(
  query,
  target,
  param,
  mzColname = "mz",
  massColname = "exactmass",
  BPPARAM = SerialParam()
)

\S4method{matchMz}{numeric,numeric,MzParam}(query, target, param, BPPARAM = SerialParam())

\S4method{matchMz}{numeric,data.frameOrSimilar,MzParam}(query, target, param, mzColname = "mz", BPPARAM = SerialParam())

\S4method{matchMz}{data.frameOrSimilar,numeric,MzParam}(query, target, param, mzColname = "mz", BPPARAM = SerialParam())

\S4method{matchMz}{data.frameOrSimilar,data.frameOrSimilar,MzParam}(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  BPPARAM = SerialParam()
)

\S4method{matchMz}{data.frameOrSimilar,data.frameOrSimilar,Mass2MzRtParam}(
  query,
  target,
  param,
  massColname = "exactmass",
  mzColname = "mz",
  rtColname = c("rt", "rt"),
  BPPARAM = SerialParam()
)

\S4method{matchMz}{data.frameOrSimilar,data.frameOrSimilar,MzRtParam}(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  rtColname = c("rt", "rt"),
  BPPARAM = SerialParam()
)

\S4method{matchMz}{SummarizedExperiment,ANY,Param}(
  query,
  target,
  param,
  mzColname = "mz",
  rtColname = "rt",
  BPPARAM = SerialParam()
)
}
\arguments{
\item{tolerance}{for any \code{param} object: \code{numeric(1)} defining the maximal
acceptable absolute difference in m/z values to consider them \emph{matching}.}

\item{ppm}{for any \code{param} object: \code{numeric(1)} defining the maximal
acceptable m/z-dependent difference (in parts-per-million) in m/z values
to consider them to be \emph{matching}.}

\item{adducts}{for \code{Mass2MzParam} or \code{Mass2MzRtParam}:
either \code{character} with the names of adducts or \code{data.frame} with the
adduct definition. This parameter is used to calculate m/z from target
compounds' masses. Custom adduct definitions can be passed to the adduct
parameter in form of a \code{data.frame}. This \code{data.frame} is expected to
have columns \code{"mass_add"} and \code{"mass_multi"} defining the \emph{additive} and
\emph{multiplicative} part of the calculation. See
\code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adducts()}} for the expected format or use
\code{MetaboCoreUtils::adductNames("positive")} and
\code{MetaboCoreUtils::adductNames("negative")} for valid adduct names.}

\item{toleranceRt}{for \code{Mass2MzRtParam} or \code{MzRtParam}: \code{numeric(1)}
defining the maximal acceptable absolute difference in retention time
values to consider them them \emph{matching}.}

\item{query}{feature table containing information on MS1 features. Can be
a \code{data.frame} (with mandatory column names \code{"mz"}) or a \code{numeric} with
the m/z values. A matching based on both m/z and retention time can be
performed when a column \code{"rt"} is present in both \code{query} and \code{target}.}

\item{target}{compound table with metabolites to compare against.}

\item{param}{parameter object defining the matching approach and containing
the settings for that approach. See description above for details.}

\item{...}{currently ignored.}

\item{BPPARAM}{parallel processing setup. See \code{BiocParallel::bpparam()} for
details.}

\item{massColname}{\code{character(1)} with the name of the column containing the
mass of compounds. Defaults to \code{massColname = "exactmass"}.}

\item{mzColname}{\code{character(1)} with the name of the column containing the
m/z values. Defaults to \code{mzColname = "mz"}.}

\item{rtColname}{\code{character(1)} with the name of the column containing the
retention times of the compounds. Defaults to \code{rtColname = "rt"}.}
}
\value{
\link{Matched} object representing the result
}
\description{
The \code{matchMz} method matches (compares) m/z values from a MS1 data table
(parameter \code{query}) with theoretical m/z values for (reference) compounds
(parameter \code{target}) considering or not also retention times. The approach
which is used to perform the comparison and additional settings for this
matching can be defined with \code{param}.

Available matching approaches and respective \code{param} objects are:
\itemize{
\item \code{Mass2MzParam}: match m/z values against reference compounds for
which the (exact) mass is known. Before matching, m/z values are calculated
from the mass of the compounds in the \emph{target} table for the specified
adducts (parameter \code{adduct} in the parameter object).
\code{query} must be a \code{data.frame} with a column containing m/z values (column
name can be specified with parameter \code{mzColname} which defaults to \code{"mz"})
or a \code{numeric} with the m/z values of the features.
If \code{target} is a \code{data.frame} it must
contain a column with the (monoisotopic) mass of the compounds (the column
name can be specified with parameter \code{massColname} which defaults to
\code{"exactmass"}) \code{Mass2MzParam}'s parameter \code{adducts} allows to define
the expected adducts (defaults to \code{adducts = "[M+H]+"} but any adducts
available in \code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adducts()}} are supported). Parameter
\code{tolerance} and \code{ppm} allow to define the maximal acceptable (constant or
m/z relative) difference between query and target m/z values.
\item \code{Mass2MzRtParam}: match m/z and retention time values against
reference compounds for which the (exact) mass \strong{and} retention time are
known. Before matching, m/z values are calculated from the mass of the
compounds in the \emph{target} table for the specified adducts (parameter
\code{adduct} in the parameter object). The retention time is considered equal
for different adducts of the same compound. \code{query} must be a \code{data.frame}
with m/z and retention time values of the features. The names of the
columns containing these information can be specified with parameters
\code{mzColname} and \code{rtColname} which default to \code{"mz"} and \code{"rt"},
respectively). \code{target} must be a \code{data.frame} with the (monoisotopic)
mass and retention time for each compound. The names of the columns
containing these information can be specified with parameters \code{massColname}
and \code{rtColname} which default to \code{"exactmass"} and \code{"rt"}, respectively.
\code{Mass2MzRtParam}'s parameter \code{adducts} allows to define the expected
adducts (defaults to \code{adducts = "[M+H]+"} but any adducts available in
\code{\link[MetaboCoreUtils:adductNames]{MetaboCoreUtils::adducts()}} are supported). Parameter \code{tolerance} and
\code{ppm} allow to define the maximal acceptable (constant or m/z relative)
difference between query and target m/z values; parameter \code{toleranceRt}
allows to specify the maximal acceptable difference between query and
target retention time values.
\item \code{MzParam}: match m/z values against reference compounds for which the m/z
is known. \code{query} must be either a \code{data.frame} with a column containing
m/z values (which name can be specified with parameter \code{mzColname},
default being \code{mzColname = "mz"}) or a \code{numeric} with the m/z values of
the features. The same holds for \code{target}. \code{MzParam} parameters
\code{tolerance} and \code{ppm} allow to define the maximal acceptable (constant or
m/z relative) difference between query and target m/z values.
\item \code{MzRtParam}: match m/z and retention time values against reference
compounds for which m/z and retention time are known. \code{query} must be a
\code{data.frame} or a \code{SummarizedExperiment}. The \code{data.frame} in one case or
the \code{SummarizedExperiment} \code{rowData} in the other must have columns
containing the m/z and retention times of the
features. The names of the respective columns can be specified with
parameters \code{mzColname} and \code{rtColname} which default to \code{"mz"} and \code{"rt"},
respectively.\code{target} must be a \code{data.frame} again with the information on
m/z and retention time.\code{MzRtParam} parameters \code{tolerance} and \code{ppm} allow
to define the maximal acceptable (constant or m/z relative) difference
between query and target m/z values; \code{MzRtParam} parameter \code{toleranceRt}
allows to specify the maximal acceptable difference between query and
target retention time values.
}
}
\examples{

library(MetaboCoreUtils)
## Create a simple "target/reference" compound table
target_df <- data.frame(
   name = c("Tryptophan", "Leucine", "Isoleucine"),
   formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
   exactmass = c(204.089878, 131.094629, 131.094629)
)

## Create a "feature" table with m/z of features. We calculate m/z for
## certain adducts of some of the compounds in the reference table.
fts <- data.frame(
    feature_id = c("FT001", "FT002", "FT003"),
    mz = c(mass2mz(204.089878, "[M+H]+"),
           mass2mz(131.094629, "[M+H]+"),
           mass2mz(204.089878, "[M+Na]+") + 1e-6))

## Define the parameters for the matching
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res

## Get the full matching result:
matchedData(res)

## We have thus matches of FT002 to two different compounds (but with the
## same mass).

## We repeat the matching requiring an exact match
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 0)
res <- matchMz(fts, target_df, parm)
res

matchedData(res)

## The last feature could thus not be matched to any compound.

## At last we use also different adduct definitions.
parm <- Mass2MzParam(
    adducts = c("[M+K]+", "[M+Li]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res

matchedData(res)

## No matches were found.

## We can also match a "feature" table with a target data.frame taking into
## account both m/z and retention time values.
target_df <- data.frame(
  name = c("Tryptophan", "Leucine", "Isoleucine"),
  formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
  exactmass = c(204.089878, 131.094629, 131.094629),
  rt = c(150, 140, 140)
)

fts <- data.frame(
  feature_id = c("FT001", "FT002", "FT003"),
  mz = c(mass2mz(204.089878, "[M+H]+"),
         mass2mz(131.094629, "[M+H]+"),
         mass2mz(204.089878, "[M+Na]+") + 1e-6),
  rt = c(150, 140, 150.1)
)

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0)

res <- matchMz(fts, target_df, parm)
res

## Get the full matching result:
matchedData(res)

## FT003 could not be matched to any compound, FT002 was matched to two
## different compounds (but with the same mass).

## We repeat the matching allowing a positive tolerance for the matches
## between rt values

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0.1)

res <- matchMz(fts, target_df, parm)
res

## Get the full matching result:
matchedData(res)

## Also FT003 was matched in this case

## It is also possible to match directly m/z values
mz1 <- c(12, 343, 23, 231)
mz2 <- mz1 + rnorm(4, sd = 0.001)

res <- matchMz(mz1, mz2, MzParam(tolerance = 0.001))

matchedData(res)
}
\seealso{
\link{matchSpectra} or \code{\link[=CompareSpectraParam]{CompareSpectraParam()}} for spectra data matching
}
\author{
Andrea Vicini, Michael Witting
}
