% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDbSource.R
\docType{class}
\name{MetFragParam,MetFragSource, validAdducts}
\alias{MetFragParam,MetFragSource, validAdducts}
\alias{MetFragParam-class}
\alias{MetFragParam}
\alias{MetFragSource}
\title{Annotation using MetFrag}
\usage{
MetFragParam(...)

MetFragSource(url = "https://msbi.ipb-halle.de/MetFrag-deNBI/api/v1/process")
}
\description{
Experimental (query) fragment spectra can be matched against the MetFrag
server using \code{matchSpectra} by providing \code{MatchFragSource()} with parameter
\code{target}. The parameters for the matching against MetFrag can be specified
using a \code{MetFragParam} object (parameter \code{param} of \code{matchSpectra}).
}
\author{
Nir Shahaf, Johannes Rainer
}
