% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMatchedSpectra.R
\name{shinyMatchedSpectra}
\alias{shinyMatchedSpectra}
\title{Verify Results from matchSpectra}
\usage{
shinyMatchedSpectra(object)
}
\arguments{
\item{object}{A non-empty instance of class \code{MatchedSpectra}.}
}
\value{
A \code{MatchedSpectra} with verified results
}
\description{
The \code{shinyMatchedSpectra()} function opens a simple shiny application
that allows to browse
}
\author{
Michael Witting
}
