% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraDbSource.R
\name{SpectraDbSource}
\alias{SpectraDbSource}
\alias{SpectraDbSource-class}
\alias{RemoteWeizMassSource}
\alias{LocalWeizMassSource}
\alias{show,SpectraDbSource-method}
\title{Compond Annotation Sources for databases supporting \code{Spectra}}
\usage{
RemoteWeizMassSource(
  IagreeToTheLicense = FALSE,
  host = character(),
  user = character(),
  pass = character(),
  dbname = character()
)

LocalWeizMassSource(IagreeToTheLicense = FALSE, dbfile = character())

\S4method{show}{SpectraDbSource}(object)
}
\arguments{
\item{IagreeToTheLicense}{\code{logical(1)} specifying whether the user accepts
to the WeizMass license agreement or not.}

\item{host}{For \code{RemoteWeizMassSource}: optional \code{character(1)} with the
name (address) of the host with the WeizMass database.}

\item{user}{For \code{RemoteWeizMassSource}: optional \code{character(1)} with the
username to access the WeizMass database.}

\item{pass}{For \code{RemoteWeizMassSource}: optional \code{character(1)} with the
password to access the WeizMass database.}

\item{dbname}{For \code{RemoteWeizMassSource}: optional \code{character(1)} with the
name of the database on the server.}

\item{dbfile}{For \code{LocalWeizMassSource}: \code{character(1)} with the (SQLite)
database file.}

\item{object}{A \code{SpectraDbSource} object.}
}
\description{
\code{SpectraDbSource} objects represent references to annotation databases that
are supported by a dedicated \link{MsBackend} backend and hence allow to return
their data as a \link{Spectra} object. A specific \code{MsBackend} has thus to be
available (and needs to be supported) by the database. Examples are
\emph{MassBank} with the
\href{https://github.com/RforMassSpectrometry/MsBackendMassbank}{MsBackendMassbank}
backend or \emph{WeizMass} with the
\href{https://github.com/RforMassSpectrometry/MsBackendWeizMass}{MsBackendWeizMass}.
Additional parameters that need to be defined are the type of SQL database
system that is used by the resource (e.g. \code{MariaDB} for MySQL/MariaDB
servers) as well as credential information allowing to connect to the
database. Connection to the database is only performed in the \code{matchSpectra}
call and the returned \link{MatchedSpectra} object contains only matching spectra
from the reference (target) database.

New \code{SpectraDbSource} objects should not be created manually, but using one
of the constructor functions listed below. Functions are available for
different annotation databases that provide also useful defaults simplifying
thus access to these resources.
\itemize{
\item \code{RemoteWeizMassSource}: connects to a \emph{remote} WeizMass (MySQL) database
hosted on a remote server. Parameter \code{user}, \code{pass}, \code{host} and \code{dbname}
are optional and allow connection to a non-standard (default) WeizMass
database (e.g. a local installation).
For WeizMass databases, users have to agree to the WeizMass license by
specifically setting parameter \code{IagreeToTheLicense} to \code{TRUE}.
\item \code{LocalWeizMassSource}: connects to a \emph{local} WeizMass database available as
a \emph{SQLite} database file. The database file has to be provided with the
\code{dbfile} parameter.
For WeizMass databases, users have to agree to the WeizMass license by
specifically setting parameter \code{IagreeToTheLicense} to \code{TRUE}.
}
}
\seealso{
Other Compound annotation sources: 
\code{\link{CompDbSource}}
}
\author{
Johannes Rainer, Nir Shachaf
}
\concept{Compound annotation sources}
